refactor: extract monitoring page controller hook
