+++ admin-ui/app/monitoring/page.tsx
@@ -1,312 +1,37 @@
 'use client';
 
-import { useEffect, useState } from 'react';
 import { PermissionGuard } from '@/components/PermissionGuard';
 import { ResponsiveLayout } from '@/components/ResponsiveLayout';
-import { useConfirm } from '@/components/ui/confirm-dialog';
-import { api } from '@/lib/api-client';
-import { isAlertSnoozed } from '@/lib/monitoring-alerts';
 import {
-  MONITORING_DEFAULT_SERIES_VISIBILITY,
-  toggleMonitoringSeriesVisibility,
-  type MonitoringMetricSeriesKey,
-  type MonitoringMetricsSeriesVisibility,
-  type MonitoringTimeRangeOption,
-} from '@/lib/monitoring-metrics';
-import {
-  MONITORING_SUBSYSTEMS,
   normalizeMonitoringHealthStatus,
 } from '@/lib/monitoring-health';
 import MonitoringFeedbackBanners from './components/MonitoringFeedbackBanners';
 import MonitoringManagementPanels from './components/MonitoringManagementPanels';
 import MonitoringMetricsSection from './components/MonitoringMetricsSection';
 import MonitoringPageHeader from './components/MonitoringPageHeader';
-import { useMonitoringDataLoader } from './use-monitoring-data-loader';
-import { useMonitoringDashboardState } from './use-monitoring-dashboard-state';
-import { useMonitoringMetricsHistory } from './use-monitoring-metrics-history';
-import { useAlertActions } from './use-alert-actions';
-import { useAlertRules } from './use-alert-rules';
-import { useMonitoringAlertState } from './use-monitoring-alert-state';
-import { useMonitoringMessages } from './use-monitoring-messages';
-import { useMonitoringNotificationState } from './use-monitoring-notification-state';
-import { useNotificationActions } from './use-notification-actions';
-import { useWatchlistActions } from './use-watchlist-actions';
-import type {
-  SystemHealthStatus,
-  SystemStatusItem,
-} from './types';
-
-const METRIC_CRITICAL_THRESHOLD = 90;
-const METRIC_WARNING_THRESHOLD = 70;
-const DEFAULT_SYSTEM_STATUS: SystemStatusItem[] = [
-  ...MONITORING_SUBSYSTEMS.map((subsystem) => ({
-    key: subsystem.key,
-    label: subsystem.label,
-    status: 'unknown' as const,
-    detail: 'Checking...',
-  })),
-];
-const MONITORING_TIME_RANGE_OPTIONS: Array<{ value: MonitoringTimeRangeOption; label: string }> = [
-  { value: '1h', label: '1h' },
-  { value: '6h', label: '6h' },
-  { value: '24h', label: '24h' },
-  { value: '7d', label: '7d' },
-  { value: '30d', label: '30d' },
-  { value: 'custom', label: 'Custom' },
-];
+import { useMonitoringPageController } from './use-monitoring-page-controller';
+import type { SystemHealthStatus } from './types';
 
 export const normalizeHealthStatus = (status?: string): SystemHealthStatus => {
   return normalizeMonitoringHealthStatus(status);
 };
 
 export default function MonitoringPage() {
-  const confirm = useConfirm();
   const {
-    metrics,
-    setMetrics,
-    watchlists,
-    setWatchlists,
-    systemStatus,
-    setSystemStatus,
-    loading,
-    setLoading,
-    lastUpdated,
-    markMonitoringDataUpdated,
-  } = useMonitoringDashboardState({
-    initialSystemStatus: DEFAULT_SYSTEM_STATUS,
-  });
-  const [seriesVisibility, setSeriesVisibility] = useState<MonitoringMetricsSeriesVisibility>(
-    MONITORING_DEFAULT_SERIES_VISIBILITY
-  );
-
-  // Notification settings
-  const {
-    notificationSettings,
-    setNotificationSettings,
-    recentNotifications,
-    setRecentNotifications,
-    setNotificationSettingsStatus,
-    canSaveNotificationSettings,
-  } = useMonitoringNotificationState();
-
-  // Stage 2: Alert rules + enhanced alert management
-  const {
-    alerts,
-    setAlerts,
-    alertsRef,
-    assignableUsers,
-    setAssignableUsers,
-    showSnoozedAlerts,
-    setShowSnoozedAlerts,
-    alertHistory,
-    setAlertHistory,
-    alertHistoryRef,
-  } = useMonitoringAlertState();
-
-  const handleManualRangeLoadSuccess = markMonitoringDataUpdated;
-
-  const {
-    metricsHistory,
-    timeRange,
-    customRangeStart,
-    customRangeEnd,
-    rangeValidationError,
-    activeRangeLabel,
-    setCustomRangeStart,
-    setCustomRangeEnd,
-    loadMetricsHistoryForRange,
-    handleSelectTimeRange,
-    handleApplyCustomTimeRange,
-  } = useMonitoringMetricsHistory({
-    apiClient: api,
-    onManualRangeLoadSuccess: handleManualRangeLoadSuccess,
-  });
-
-  const handleToggleSeries = (seriesKey: MonitoringMetricSeriesKey) => {
-    setSeriesVisibility((prev) => toggleMonitoringSeriesVisibility(prev, seriesKey));
-  };
-
-  const {
-    error,
-    setError,
-    success,
-    setSuccess,
-  } = useMonitoringMessages();
-
-  const { loadData } = useMonitoringDataLoader({
-    alertsRef,
-    alertHistoryRef,
-    customRangeStart,
-    customRangeEnd,
-    timeRange,
-    loadMetricsHistoryForRange,
-    markMonitoringDataUpdated,
-    setLoading,
-    setError,
-    setNotificationSettingsStatus,
-    setNotificationSettings,
-    setRecentNotifications,
-    setMetrics,
-    setWatchlists,
-    setAlerts,
-    setAlertHistory,
-    setAssignableUsers,
-    setSystemStatus,
-    metricWarningThreshold: METRIC_WARNING_THRESHOLD,
-    metricCriticalThreshold: METRIC_CRITICAL_THRESHOLD,
-  });
-
-  const {
-    showCreateWatchlist,
-    setShowCreateWatchlist,
-    newWatchlist,
-    setNewWatchlist,
-    deletingWatchlistId,
-    handleCreateWatchlist,
-    handleDeleteWatchlist,
-  } = useWatchlistActions({
-    apiClient: api,
-    confirm,
-    setError,
-    setSuccess,
-    onReloadRequested: loadData,
-  });
-
-  const {
-    handleAcknowledgeAlert,
-    handleDismissAlert,
-    handleAssignAlert,
-    handleSnoozeAlert,
-    handleEscalateAlert,
-  } = useAlertActions({
-    apiClient: api,
-    confirm,
-    setAlerts,
-    setAlertHistory,
-    setError,
-    setSuccess,
-    onReloadRequested: loadData,
-    assignableUsers,
-  });
-
-  const {
-    notificationsSaving,
-    handleSaveNotificationSettings,
-    handleTestNotification,
-  } = useNotificationActions({
-    apiClient: api,
-    canSave: canSaveNotificationSettings,
-    setNotificationSettings,
-    setRecentNotifications,
-    setError,
-    setSuccess,
-  });
-
-  const {
-    alertRules,
-    alertRuleDraft,
-    alertRuleValidationErrors,
-    alertRulesSaving,
-    handleAlertRuleDraftChange,
-    handleCreateAlertRule,
-    handleDeleteAlertRule,
-  } = useAlertRules({
-    setSuccess,
-  });
-
-  useEffect(() => {
-    void loadData();
-  }, [loadData]);
-
-  const activeAlerts = alerts.filter((alert) => !alert.acknowledged && !isAlertSnoozed(alert));
+    headerProps,
+    feedbackBannersProps,
+    metricsSectionProps,
+    managementPanelsProps,
+  } = useMonitoringPageController();
 
   return (
     <PermissionGuard variant="route" requireAuth role="admin">
       <ResponsiveLayout>
         <div className="p-4 lg:p-8">
-          <MonitoringPageHeader
-            lastUpdated={lastUpdated}
-            loading={loading}
-            onRefresh={loadData}
-          />
-
-          <MonitoringFeedbackBanners
-            error={error}
-            success={success}
-            activeAlertsCount={activeAlerts.length}
-          />
-
-          <MonitoringMetricsSection
-            timeRangeControlsProps={{
-              options: MONITORING_TIME_RANGE_OPTIONS,
-              timeRange,
-              customRangeStart,
-              customRangeEnd,
-              rangeValidationError,
-              onSelectTimeRange: handleSelectTimeRange,
-              onCustomRangeStartChange: setCustomRangeStart,
-              onCustomRangeEndChange: setCustomRangeEnd,
-              onApplyCustomRange: handleApplyCustomTimeRange,
-            }}
-            metricsChartProps={{
-              metricsHistory,
-              rangeLabel: activeRangeLabel,
-              seriesVisibility,
-              onToggleSeries: handleToggleSeries,
-            }}
-            metricsGridProps={{
-              metrics,
-              loading,
-            }}
-          />
-
-          <MonitoringManagementPanels
-            alertRulesPanelProps={{
-              rules: alertRules,
-              draft: alertRuleDraft,
-              errors: alertRuleValidationErrors,
-              saving: alertRulesSaving,
-              onDraftChange: handleAlertRuleDraftChange,
-              onCreateRule: handleCreateAlertRule,
-              onDeleteRule: handleDeleteAlertRule,
-            }}
-            alertsPanelProps={{
-              alerts,
-              history: alertHistory,
-              showSnoozed: showSnoozedAlerts,
-              assignableUsers,
-              loading,
-              onToggleShowSnoozed: () => setShowSnoozedAlerts((prev) => !prev),
-              onAcknowledge: handleAcknowledgeAlert,
-              onDismiss: handleDismissAlert,
-              onAssign: handleAssignAlert,
-              onSnooze: handleSnoozeAlert,
-              onEscalate: handleEscalateAlert,
-            }}
-            watchlistsPanelProps={{
-              watchlists,
-              loading,
-              showCreateWatchlist,
-              setShowCreateWatchlist,
-              newWatchlist,
-              setNewWatchlist,
-              onCreate: handleCreateWatchlist,
-              onDelete: handleDeleteWatchlist,
-              deletingWatchlistId,
-            }}
-            notificationsPanelProps={{
-              settings: notificationSettings,
-              recentNotifications,
-              loading,
-              saving: notificationsSaving,
-              canSave: canSaveNotificationSettings,
-              onSave: handleSaveNotificationSettings,
-              onTest: handleTestNotification,
-            }}
-            systemStatusPanelProps={{
-              systemStatus,
-            }}
-          />
+          <MonitoringPageHeader {...headerProps} />
+          <MonitoringFeedbackBanners {...feedbackBannersProps} />
+          <MonitoringMetricsSection {...metricsSectionProps} />
+          <MonitoringManagementPanels {...managementPanelsProps} />
         </div>
       </ResponsiveLayout>
     </PermissionGuard>
